requirements.txt
# Core dependencies
requests>=2.31.0
pyyaml>=6.0.1
jinja2>=3.1.2
python-dotenv>=1.0.0

# CLI
colorama>=0.4.6

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1

# Code quality
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

